+++ IBS/DataAccess/T18CallDetail.cs
@@ -9,7 +9,7 @@ public partial class T18CallDetail
 
     public DateTime CallRecvDt { get; set; }
 
-    public short CallSno { get; set; }
+    public int CallSno { get; set; }
 
     public int ItemSrnoPo { get; set; }
